[feat] 계정 권한 필터 추가 - 선물 바구니 삭제, 수정, 조회, 상품 추가, 삭제

+++ src/main/java/clov3r/oneit_server/controller/GiftboxController.java
@@ -46,16 +46,12 @@ public class GiftboxController {
       @Parameter(hidden = true) @Auth Long userIdx
   ) {
     try {
+      // request validation
+      if (request.getName() == null || request.getDeadline() == null || userIdx == null) {
+        return new BaseResponse<>(REQUEST_ERROR);
+      }
       if (!userRepository.existsUser(userIdx)) {
         return new BaseResponse<>(USER_NOT_FOUND);
-      } else {
-        request.setCreatedUserIdx(userIdx);
-      }
-
-      // request validation
-      if (request.getName() == null || request.getDeadline() == null
-          || request.getCreatedUserIdx() == null) {
-        return new BaseResponse<>(REQUEST_ERROR);
       }
       if (request.getDeadline().isBefore(LocalDateTime.now().toLocalDate())) {
         return new BaseResponse<>(DATE_BEFORE_NOW);
@@ -64,7 +60,11 @@ public class GiftboxController {
           && !request.getAccessStatus().equals(AccessStatus.PRIVATE.toString())) {
         return new BaseResponse<>(INVALID_ACCESS_STATUS);
       }
-      Long giftboxIdx = giftboxService.createGiftbox(request);
+
+      // create Gift box
+      Long giftboxIdx = giftboxService.createGiftbox(request, userIdx);
+
+      // upload image if exists
       if (image != null) {
         String imageUrl = s3Service.upload(image, "giftbox-profile");
         giftboxService.updateGiftboxImageUrl(giftboxIdx, imageUrl);
@@ -80,15 +80,29 @@ public class GiftboxController {
   @Operation(summary = "Giftbox 상세 조회", description = "선물 바구니의 idx로 선물 바구니 조회, 삭제한 선물 바구니는 조회되지 않음")
   @GetMapping("/api/v1/giftbox/{giftboxIdx}")
   public BaseResponse<GiftboxDTO> getGiftboxByIdx(
-      @PathVariable("giftboxIdx") Long giftboxIdx)
+      @PathVariable("giftboxIdx") Long giftboxIdx,
+      @Parameter(hidden = true) @Auth Long userIdx
+  )
   {
-
+    // request validation
+    if (giftboxIdx == null || userIdx == null) {
+      return new BaseResponse<>(REQUEST_ERROR);
+    }
     if (!giftboxRepository.existsById(giftboxIdx)) {
       return new BaseResponse<>(GIFTBOX_NOT_FOUND);
     }
+    if (!userRepository.existsUser(userIdx)) {
+      return new BaseResponse<>(USER_NOT_FOUND);
+    }
 
+    // get giftbox
     try {
       Giftbox giftbox = giftboxRepository.findById(giftboxIdx);
+
+      if (giftbox.getAccessStatus().equals(AccessStatus.PRIVATE) && !giftboxRepository.isParticipantOfGiftbox(userIdx, giftboxIdx)) {
+        return new BaseResponse<>(NOT_PARTICIPANT_OF_GIFTBOX);  // 선물 바구니가 PRIVATE 일 경우, 해당 선물 바구니의 참여자만 조회 가능함
+      }
+
       GiftboxDTO giftboxDTO = new GiftboxDTO(
           giftbox.getIdx(),
           giftbox.getName(),
@@ -105,7 +119,7 @@ public class GiftboxController {
   }
 
   @Tag(name = "Giftbox API", description = "Giftbox CRUD API 목록")
-  @Operation(summary = "Giftbox 목록 조회", description = "모든 선물 바구니 조회, 삭제한 선물 바구니는 조회되지 않음")
+  @Operation(summary = "Giftbox 목록 조회(추후 삭제 예정)", description = "모든 선물 바구니 조회, 삭제한 선물 바구니는 조회되지 않음")
   @GetMapping("/api/v1/giftbox/all")
   public BaseResponse<List<GiftboxDTO>> getGiftboxList() {
     try {
@@ -123,14 +137,14 @@ public class GiftboxController {
   @Operation(summary = "해당 유저의 Giftbox 목록 조회", description = "해당 유저가 소유한 모든 선물 바구니 조회, 삭제한 선물 바구니는 조회되지 않음")
   @GetMapping("/api/v1/giftbox")
   public BaseResponse<List<GiftboxDTO>> getGiftboxList(
-      @Parameter(hidden = true)
-      @RequestParam("userIdx") Long userIdx
+      @Parameter(hidden = true) @Auth Long userIdx
   ) {
-
+    // request validation
     if (!userRepository.existsUser(userIdx)) {
       return new BaseResponse<>(USER_NOT_FOUND);
     }
 
+    // get giftbox list of the user
     try {
       List<Giftbox> giftboxList = giftboxRepository.findGiftboxOfUser(userIdx);
       List<GiftboxDTO> giftboxDTOList = giftboxList.stream()
@@ -145,10 +159,22 @@ public class GiftboxController {
   @Tag(name = "Giftbox API", description = "Giftbox CRUD API 목록")
   @Operation(summary = "Giftbox 삭제", description = "선물 바구니의 idx로 선물 바구니 삭제")
   @DeleteMapping("/api/v1/giftbox/{giftboxIdx}")
-  public BaseResponse<String> deleteGiftboxByIdx(@PathVariable("giftboxIdx") Long giftboxIdx) {
+  public BaseResponse<String> deleteGiftboxByIdx(
+      @PathVariable("giftboxIdx") Long giftboxIdx,
+      @Parameter(hidden = true) @Auth Long userIdx
+  ) {
+    // request validation
     if (giftboxRepository.findById(giftboxIdx) == null) {
       return new BaseResponse<>(GIFTBOX_NOT_FOUND);
     }
+    if (!userRepository.existsUser(userIdx)) {
+      return new BaseResponse<>(USER_NOT_FOUND);
+    }
+    if (!giftboxRepository.isManagerOfGiftbox(userIdx, giftboxIdx)) {
+      return new BaseResponse<>(NOT_MANAGER_OF_GIFTBOX);  // 해당 선물 바구니의 관리자만 삭제 가능함
+    }
+
+    // delete giftbox
     try {
       giftboxRepository.deleteById(giftboxIdx);
       return new BaseResponse<>(giftboxIdx + "번 선물 바구니가 삭제되었습니다.");
@@ -163,19 +189,22 @@ public class GiftboxController {
   public BaseResponse<Long> updateGiftbox(
       @PathVariable("giftboxIdx") Long giftboxIdx,
       @RequestPart("request") PostGiftboxRequest request,
-      @RequestPart(value = "image", required = false) MultipartFile image)
-  {
+      @RequestPart(value = "image", required = false) MultipartFile image,
+      @Parameter(hidden = true) @Auth Long userIdx
+  ) {
     try {
       // request validation
       if (giftboxRepository.findById(giftboxIdx) == null) {
         return new BaseResponse<>(GIFTBOX_NOT_FOUND);
       }
-      if (request.getName() == null || request.getDeadline() == null
-          || request.getCreatedUserIdx() == null) {
+      if (request.getName() == null || request.getDeadline() == null || userIdx == null) {
         return new BaseResponse<>(REQUEST_ERROR);
       }
-      if (!userRepository.existsUser(request.getCreatedUserIdx())) {
+      if (!userRepository.existsUser(userIdx)) {
         return new BaseResponse<>(USER_NOT_FOUND);
+      }
+      if (!giftboxRepository.isManagerOfGiftbox(userIdx, giftboxIdx)) {
+        return new BaseResponse<>(NOT_MANAGER_OF_GIFTBOX);  // 해당 선물 바구니의 관리자만 수정 가능함
       }
       if (request.getDeadline().isBefore(LocalDateTime.now().toLocalDate())) {
         return new BaseResponse<>(DATE_BEFORE_NOW);
@@ -184,7 +213,11 @@ public class GiftboxController {
           && !request.getAccessStatus().equals(AccessStatus.PRIVATE.toString())) {
         return new BaseResponse<>(INVALID_ACCESS_STATUS);
       }
+
+      // update giftbox
       giftboxService.updateGiftbox(giftboxIdx, request);
+
+      // upload image if exists
       if (image != null) {
         String imageUrl = s3Service.upload(image, "giftbox-profile");
         giftboxService.updateGiftboxImageUrl(giftboxIdx, imageUrl);
@@ -196,24 +229,30 @@ public class GiftboxController {
     }
   }
 
-  // 상품 위시리스트에 상품 리스트를 추가하는 API
+  // 선물 바구니에 상품 리스트를 추가하는 API
   @Tag(name = "Giftbox 상품 API", description = "Giftbox 상품 CRUD API 목록")
   @Operation(summary = "Giftbox 상품 추가", description = "선물 바구니의 idx 리스트로 상품 추가")
   @PostMapping("/api/v1/giftbox/{giftboxIdx}/products")
   @Transactional
   public BaseResponse<String> addProductToGiftbox(
       @PathVariable("giftboxIdx") Long giftboxIdx,
-      @RequestBody List<Long> productIdxList) {
+      @RequestBody List<Long> productIdxList,
+      @Parameter(hidden = true) @Auth Long userIdx
+  ) {
+    // request validation
     if (giftboxRepository.findById(giftboxIdx) == null) {
       return new BaseResponse<>(GIFTBOX_NOT_FOUND);
     }
-
     for(Long productIdx : productIdxList) {
       if (!productRepository.existsProduct(productIdx)) {
         return new BaseResponse<>(PRODUCT_NOT_FOUND);
       }
     }
+    if (!giftboxRepository.isParticipantOfGiftbox(userIdx, giftboxIdx)) {
+      return new BaseResponse<>(NOT_PARTICIPANT_OF_GIFTBOX);  // 해당 선물 바구니의 참여자만 상품 추가 가능함
+    }
 
+    // add product to giftbox
     try {
       for (Long productIdx : productIdxList) {
         giftboxService.addProductToGiftbox(giftboxIdx, productIdx);
@@ -232,9 +271,12 @@ public class GiftboxController {
   public BaseResponse<List<ProductSummaryDTO>> getProductOfGiftbox(
       @PathVariable("giftboxIdx") Long giftboxIdx
   ) {
+    // request validation
     if (giftboxRepository.findById(giftboxIdx) == null) {
       return new BaseResponse<>(GIFTBOX_NOT_FOUND);
     }
+
+    // get product list of the giftbox
     List<Product> productList = giftboxRepository.findProductOfGiftbox(giftboxIdx);
     List<ProductSummaryDTO> productSummaryDTOList = productList.stream()
         .map(ProductSummaryDTO::new)
@@ -248,18 +290,29 @@ public class GiftboxController {
   @DeleteMapping("/api/v1/giftbox/{giftboxIdx}/products")
   public BaseResponse<String> deleteProductsOfGiftbox(
       @PathVariable("giftboxIdx") Long giftboxIdx,
-      @RequestBody List<Long> productIdxList
+      @RequestBody List<Long> productIdxList,
+      @Parameter(hidden = true) @Auth Long userIdx
   ) {
+    // request validation
+    if (giftboxIdx == null || productIdxList == null || userIdx == null) {
+      return new BaseResponse<>(REQUEST_ERROR);
+    }
     if (giftboxRepository.findById(giftboxIdx) == null) {
       return new BaseResponse<>(GIFTBOX_NOT_FOUND);
     }
-
+    if (!userRepository.existsUser(userIdx)) {
+      return new BaseResponse<>(USER_NOT_FOUND);
+    }
+    if (!giftboxRepository.isParticipantOfGiftbox(userIdx, giftboxIdx)) {
+      return new BaseResponse<>(NOT_PARTICIPANT_OF_GIFTBOX);  // 해당 선물 바구니의 참여자만 상품 삭제 가능함
+    }
     for(Long productIdx : productIdxList) {
       if (!productRepository.existsProduct(productIdx)) {
         return new BaseResponse<>(PRODUCT_NOT_FOUND);
       }
     }
 
+    // delete products of the giftbox
     for (Long productIdx : productIdxList) {
       try {
         giftboxRepository.deleteProductOfGiftbox(giftboxIdx, productIdx);
